Fix more size_t truncations in fxjs
-- use checked_cast<> where it can't be avoided.
-- re-write two loops for simplicity.
-- use npos to avoid signed/unsigned comparisons.

Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/87030
Commit-Queue: Tom Sepez <tsepez@chromium.org>

+++ fxjs/cfx_v8.cpp
@@ -46,17 +46,17 @@ v8::Local<v8::Object> CFX_V8::NewObject() {
 }
 
 void CFX_V8::PutArrayElement(v8::Local<v8::Array> pArray,
-                             unsigned index,
+                             size_t index,
                              v8::Local<v8::Value> pValue) {
   fxv8::ReentrantPutArrayElementHelper(GetIsolate(), pArray, index, pValue);
 }
 
 v8::Local<v8::Value> CFX_V8::GetArrayElement(v8::Local<v8::Array> pArray,
-                                             unsigned index) {
+                                             size_t index) {
   return fxv8::ReentrantGetArrayElementHelper(GetIsolate(), pArray, index);
 }
 
-unsigned CFX_V8::GetArrayLength(v8::Local<v8::Array> pArray) {
+size_t CFX_V8::GetArrayLength(v8::Local<v8::Array> pArray) {
   return fxv8::GetArrayLengthHelper(pArray);
 }
